fix(v6): address PR #50 review (Round 10) & lazy-loading hardening
- app: Move src.ui.steps import into fragment to enable true lazy-loading of heavy modules.
- build: Refactor exclusion list generation to use packaging.requirements for robust dependency parsing.
- data_loader: Explicitly cast coerced score columns to float for consistent dtypes.
- solver: Replace 'pass' with '...' in abstract methods to satisfy PIE790 and modern standards.
- services: Move solver_interface import to OptimizationService.run for lazy-loading.
- tests: Tighten type hints in conftest.py and relax determinism status checks to include FEASIBLE solutions.
- tests: Refactor infra tests for exact basename matching and implement deterministic button mocking.
- readme: Update Groupers documentation to accurately reflect their nature as weighted soft preferences.
- pre_ci: Harden unit test phase by explicitly enforcing the 95% coverage threshold.

+++ src/core/data_loader.py
@@ -181,7 +181,7 @@ def _process_data_service(df: pd.DataFrame, filepath: str) -> list[dict] | None:
             df[col] = df[col].fillna("").astype(str)
 
     for col in score_cols:
-        df[col] = pd.to_numeric(df[col], errors="coerce").fillna(0)
+        df[col] = pd.to_numeric(df[col], errors="coerce").fillna(0.0).astype(float)
 
     records = df.to_dict("records")
     if not records:
